Made functions.py file that contains all scraping functions to be used in app.py.
